requirements.txt
# 텔레그램 봇
python-telegram-bot==20.7

# 웹 크롤링
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
scrapy==2.11.0

# 스케줄링
APScheduler==3.10.4

# 한국어 NLP
konlpy==0.6.0

# 감정 분석 및 ML
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# 데이터베이스
psycopg2-binary==2.9.9
supabase==2.1.0

# 캐싱
redis==5.0.1

# 비동기 처리
aiohttp==3.9.1
asyncio-mqtt==0.13.0

# 환경 변수 관리
python-dotenv==1.0.0

# 로깅
loguru==0.7.2

# 유틸리티
python-dateutil==2.8.2
pytz==2023.3

# 개발 도구
pytest==7.4.3
black==23.11.0
flake8==6.1.0 